reducer을 store.js에서 분리하기
